QuestEditor:     Добавлен поиск по квестам, ищется NPCReputation     Поиск ошибок, исправлена ошибка поиска, если диалог даёт квест и завершает квест и не закрывается     Исправление ошибки выбора предметов в награде/штрафе     В TreeNode добавлены ссылки на квесты, которые открываются при помощи Награды другим квестом     Убрана возможность добавить NPC(Теперь NPC добавляется в NPCEditor)     Добавлена возможность добавить NPC через NPCEditor(если запущены оба)     Добавлен тип квеста 100 - проверка всех дочерних квестов     Убран тип квеста 50 - игра против режиссёра(старые сценарии)

+++ StalkerOnlineQuesterEditor/GUIClasses/NodeDragHandler.cs
@@ -61,7 +61,7 @@ namespace StalkerOnlineQuesterEditor
 
             int rectId;
             if (mainForm.RectManager.CheckIfRect(e.PickedNode.Tag, out rectId))
-                mainForm.RectManager.ChangeCoordinates(mainForm.GetCurrentNPC(), rectId, (int) x, (int) y);
+                mainForm.RectManager.ChangeCoordinates(mainForm.GetCurrentNPC(), rectId, x, y);
         }
 
         //! Клик мыши по узлу диалога или прямоугольнику - выделяем диалог и его потомков цветом, либо выделяем прямугольник
